split exec package
