+++ pkg/kt/exec/run.go
@@ -1,4 +1,4 @@
-package util
+package exec
 
 import (
 	"bytes"
